# api.yaml
swagger: "2.0"
info:
  title: qviple.com API
  description: This is qviple.com API Documentation
  version: one
host: 18.205.27.165
basePath: /api
...
schemes: 
  - http


#Return Users From get request
paths:
...
  /api/v1/message:
    post:
      tags:
        - Qviple Socket (one to one chat)
      summary: Create message (socket)
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in : body
          name: body
          schema:
            type: object
            properties:
              content:
                type: string
                description: Enter messaging content
              chatId:
                type: string
                description: Enter current chat Id
              replyContent:
                type: string
                description: Enter Reply Content when click on reply of particular messages
              replyIndex:
                type: string
                description: Enter Reply Content Index when click on reply of particular messages
      responses:
        "200":
          description: OK
  /api/v1/message/document:
    post:
      tags:
        - Qviple Socket (one to one chat)
      summary: Create message with documents (Image, Gif, Pdf document, Doc) (socket)
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: formData
          name: file
          type: file
          required: true
          description: upload your message document
        - in: formData
          name: content
          type: string
          required: true
          description: enter messaging content
        - in: formData
          name: chatId
          type: string
          required: true
          description: Enter current chat Id
        - in: formData
          name: replyContent
          type: string
          description: Enter Reply Content when click on reply of particular messages
        - in: formData
          name: replyIndex
          type: string
          description: Enter Reply Content Index when click on reply of particular messages
      responses:
        "200":
          description: OK
...